fix: problems from the second delivery
